Added comments
I read through the code and put as many comments as I could although I
still don't fully understand the code.  Looks pretty sweet so far
though!
I'm not 100% sure what the purpose of "Console.java" is...
I'm still having problems interacting with the files on my computer.
BUT I figured out I can copy the files to my computer, edit it with
Notepad, and then upload it.  To use the IDE I can make a new project
and copy the .java files and then interact with them.