Make changes to chatresponses
I didn't understand your method that created the five chatresponses so I
commented it out and created my own.  If we can get yours to work we can
comment mine out.  The STUDENTLIST class is giving me problems.  Right
now it says:
//Getting an error at "students.add(newstudent);
for (String name: studentRoster){
String[] splitname = name.split(" ");
Student newstudent = new Student(splitname[0], splitname[1], 0, chats);
students.add(newstudent); }

At "students.add(newstudent) the compiler is giving me an error.  I
don't understand this method or its purpose so you may have to fix this
one.

I made some notes of what each of the classes do and how they work with
each other for my own benefit:
Lab2 - Does nothing, is our main method.
Console - Does nothing right now.  Will be used to have our completed
console stuff.
ChatResponses - Generates the list of 5 responses to put into our
Student class.
Student - Creates the students.
Group - Appears to contain an array or list of the students in the
groups.  Additionally the varName.getChat(); method appears to output
the chat of the entire group.
StudentList - Just using the constructor for StudentList() appears to
create the entire list of every student in the class, then puts them
into 6 groups.  varName.getGroups() returns an array/list of the groups.
Each group contains multiple students.

It appears to me that if we can wire out the bugs to run this on the
console mainly what we'll do is create a StudentList VarName = new
StudentList() which will do most of the work.  Then from that point
we'll need to call methods to get the chats to work.  Perhaps the
VarName.GetChat() in the Group Class will allow the students to talk?

I'll keep working on it but the bugs are confusing since I'm not 100%
sure what the methods are supposed to do.